Do partition assignment using pod name and total block builders and partitions
This move is in order to make partitions stable in a block builder to most extent.
This is not aimed to be the final design, but a temporary stop-gap to help test block builders.
Note: If we plan to keep it this way, we need to make identifying the number of block builders and partitions automatic, using some discovery mechanism.

Signed-off-by: Ganesh Vernekar <ganeshvern@gmail.com>